requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
defusedxml==0.7.1
et-xmlfile==1.1.0
exceptiongroup==1.1.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
igramscraper==0.3.5
iniconfig==2.0.0
instaloader==4.9.5
jedi==0.18.2
lxml==4.9.2
msgpack==1.0.4
nodeenv==1.7.0
numpy==1.24.1
odfpy==1.4.1
openpyxl==3.0.10
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathlib==1.0.1
pluggy==1.0.0
pynvim==0.4.3
pyright==1.1.290
PySimpleGUI==4.60.4
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==3.0.2
pytz==2022.7
requests==2.28.2
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
text-unidecode==1.2
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
webdriver-manager==3.8.5
wsproto==1.2.0
yapf==0.32.0
